This one is a bit counter-intuitive.

Before/after:
```
chia/table/parallel     time:   [108.23 ms 109.01 ms 109.63 ms]
chia/table/parallel     time:   [119.45 ms 123.55 ms 127.52 ms]
```

Limited to 4 efficiency cores before/after:
```
chia/table/parallel     time:   [487.01 ms 489.29 ms 492.04 ms]
chia/table/parallel     time:   [531.40 ms 536.19 ms 541.15 ms]
```

So you might be wondering, why on Earth would I send a PR that makes things slower?

2 parallel benches before/after:
```
chia/table/parallel     time:   [189.94 ms 192.78 ms 195.86 ms]
chia/table/parallel     time:   [178.51 ms 185.22 ms 191.66 ms]
```

4 parallel benches before/after:
```
chia/table/parallel     time:   [451.34 ms 458.56 ms 465.66 ms]
chia/table/parallel     time:   [334.04 ms 338.05 ms 342.08 ms]
```

As you can see, old code is significantly slower with concurrency 2 or 4, new version improves performance down to 85ms per table when 4 tables are constructed concurrently. This is because updated implementation is doing strictly less work for each table, such that when other tables are processed concurrently they have more resources available to them, but every individual table generation is a bit slower.

With the performance level achieved already for single table creation scalability that improves plotting speeds seems much more important.

I plan to submit changes later that will increase default concurrency on farmer to take advantage of these improvements and maybe introduce optimizations of similar nature (and hopefully https://github.com/supranational/blst/pull/203) to push plotting performance further than it ever was. It should also help to further reduce replotting impact on normal usage of the computer.

Please let me know if you think this tradeoff is not worth it or have any other concerns.

P.S. As a bonus parallel PoSpace is now supported in no-std environment, not that I expect anyone to take advantage of that :upside_down_face: 

### Code contributor checklist:
* [x] I have read, understood and followed [contributing guide](https://github.com/subspace/subspace/blob/main/CONTRIBUTING.md)